raspberry/requirements.txt
# MQTT Client
paho-mqtt>=1.6.1

# Async framework
asyncio-mqtt>=0.16.1

# Web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0

# Database support
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Schema validation
jsonschema>=4.20.0
pyyaml>=6.0.1
pydantic>=2.5.0

# Data processing
pandas>=2.1.3
numpy>=1.26.2

# Compression
zstandard>=0.22.0
lz4>=4.3.2

# Encryption
cryptography>=41.0.7

# Monitoring
prometheus-client>=0.19.0

# InfluxDB
influxdb-client>=1.38.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.1

# Templating (for web interface)
jinja2>=3.1.2

# Configuration
python-dotenv>=1.0.0
click>=8.1.7

# Utilities
aiofiles>=23.2.1
python-dateutil>=2.8.2
pytz>=2023.3

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
